+++ src/App.tsx
@@ -1,4 +1,4 @@
-import LoginPage from "./components/pages/LoginPage/LoginPage";
+import LoginPage from "./components/pages/loginPage/LoginPage";
 
 function App() {
